requirements.txt
streamlit
python-docx
PyPDF2
openai
langchain
pdfkit